Add files via upload
Initial exploration of datasets for the research. I am considering either using a more powerful computer (in terms of RAM) or using R, which seems to be more efficient at dealing with larger datasets. The problem is that Dask gives scalability to the program, but it does not provide all the features of pandas.